[frontend] Show significance chips in the run detail view
This doesn't quite do what you'd expect as it shows the change for
*all* dimensions. The backend would need to provide the relevant ones
separately or mark them as such.
